Invite bugfixes
1. Inviting a user via their profile page did not work because the call
to ‘inviteUser’ did not include the userId key. Also it used to
previously report user has been invited regardless of success or
failure.
2. Update to user in inviting user object in inviteUser method had a
double $inc modifier meaning that inviteCount would never actually be
decreased.
3. Added a uninviteUser method. Previously uninviting a user from the
all-users admin page would just update that user’s isInvited flag to
false. This was a problem because if that user was to be invited again
it wouldn’t work because a invite would still exist for them in the
Invites collection. This new method removes them both.