don't show future posts in daily view
currently posts scheduled for the future still show up in the daily view for regular users, which is undesirable. This checks to see if the current time is before the end of the day in question, and if so restrains the query to cap itself to only posts before the current moment.