Update graphql.js
this is a followup on the question i asked on slack. my understanding is that graphql gets confused when we give it `$` or `.`. Since any field that uses `$` in their name will also use `.`, I thought we could just amend the rule so that not only `Array`s get skipped, but also `Object`s.