Allow guests to create document as a default
Hi, using when using the default mutations, you can't allow a guest user to create a document, even if you allowed it for `guests` when setting up permissions. This is because a undefined `user` will always trigger a `false` during the "new" check.

I think it is safe to remove this, because anyway you have to manually tell who can create document (or so I think, maybe I'm  wrong?) so you can't allow guests to create documents by mistake.