Fix datatable bug when sorting a column
see https://github.com/VulcanJS/Vulcan/issues/2090#issuecomment-433860782

Previously : when you clicked on a sortable column, the results of the fired query would be null, and the next polling would make it right. You had a long time with no results displayed.

Now : The query fired when you click to sort a column returns the right result, with the props updating correctly, so no need to wait for the polling to update the results.

See #2090 for an in-depth explanation