Wrap ErrorCatcher with `withSiteData`
This way you can add multiple error catchers in your react tree. My use case is adding one in the Modal component, so that when you add a buggy thing in the modal it's caught here and the user can exit the modal and the rest of the app is still rendered.