Throwing from a validation function fails validation
As written, `runCallbacks` has the very unfortunate property that if a
callback throws an exception, it gets logged to stderr but is otherwise
ignored. This is especially problematic in the context of validation
functions, where throwing an exception is a customary way to signal that
validation has failed, and where an unexpected error like
dereferencing null *definitely* means validation should fail.

This gives `runCallbacks` an `ignoreExceptions` argument, which defaults
to true (because back-compat), and which is false for validation
functions in particular. This is not a complete solution, since there's
also validation happening in create and update callbacks, but those have
to be approached gradually.